After importing powerups, Lumos looks for Setup scripts of type ILumosSetup and runs them.

+++ Assets/Test.cs
@@ -6,23 +6,26 @@ public class Test : MonoBehaviour
 {
 	void Awake()
 	{
+		
+		
 		Lumos.debug = true;
 		
 		//LumosEvents.Record("what_up", 0, true, "levels");
 		//LumosEvents.Record("event_test", Time.time, true, "loading");
 		//LumosEvents.Record("event_test_11", 7, true, "loading");
 	}
+	
 
 	void OnGUI()
 	{
 		
-		/*if (GUILayout.Button("Delete prefs")) {
+		if (GUILayout.Button("Delete prefs")) {
 			EditorPrefs.DeleteKey("lumos-installed-packages");
 			EditorPrefs.DeleteKey("lumos-latest-packages");
 			EditorPrefs.SetBool("lumos-installing", false);
 			EditorPrefs.DeleteKey("lumos-install-queue");
 		}
-		*/
+		
 	/*	if (GUILayout.Button("Show achievements")) {
 			Social.ShowAchievementsUI();
 		}
